# avoid touching MDC for context.log.trace in reliable delivery