--- a/akka-bench-jmh/src/main/scala/akka/actor/typed/delivery/ReliableDeliveryBenchmark.scala
+++ b/akka-bench-jmh/src/main/scala/akka/actor/typed/delivery/ReliableDeliveryBenchmark.scala
@@ -60,13 +60,15 @@ object Consumer {
 
   def apply(consumerController: ActorRef[ConsumerController.Command[Command]]): Behavior[Command] = {
     Behaviors.setup { context =>
+      val traceEnabled = context.log.isTraceEnabled
       val deliveryAdapter =
         context.messageAdapter[ConsumerController.Delivery[Command]](WrappedDelivery(_))
       consumerController ! ConsumerController.Start(deliveryAdapter)
 
       Behaviors.receiveMessagePartial {
         case WrappedDelivery(d @ ConsumerController.Delivery(_, confirmTo)) =>
-          context.log.trace("Processed {}", d.seqNr)
+          if (traceEnabled)
+            context.log.trace("Processed {}", d.seqNr)
           confirmTo ! ConsumerController.Confirmed
           Behaviors.same
       }